alerts: deploy alerts function to cloudrun
Moves the alerts function out of supabase edge functions and into cloudrun.
There's no functional changes to the alerts function itself, apart from just
moving the files and changing how it's deployed.
