+++ .github/workflows/deploy-alert-api.yaml
@@ -1,0 +1,44 @@
+name: Deploy alerts
+
+on:
+  workflow_dispatch: {}
+
+jobs:
+  build:
+    runs-on: ubuntu-24.04
+    permissions:
+      # Permissions required of the Github token in order for
+      # federated identity and authorization to work.
+      contents: read
+      id-token: write
+    steps:
+      - name: Checkout
+        uses: actions/checkout@v4
+        with:
+          submodules: true
+          lfs: true
+
+      - uses: supabase/setup-cli@v1
+      - run: supabase start
+
+      - name: Authenticate with GCP Workload Identity Federation
+        uses: google-github-actions/auth@v2
+        with:
+          service_account: cd-github-actions@estuary-control.iam.gserviceaccount.com
+          workload_identity_provider: projects/1084703453822/locations/global/workloadIdentityPools/github-actions/providers/github-actions-provider
+
+      - name: Update Cloud Run service `alerts`
+        uses: google-github-actions/deploy-cloudrun@v2
+        with:
+          service: alerts
+          project_id: estuary-control
+          region: us-central1
+          source: edge-functions/alerts
+          timeout: 1m
+          secrets: |-
+            RESEND_API_KEY=ALERTS_RESEND_API_KEY:latest
+            ALERT_EMAIL_FUNCTION_SECRET=ALERT_EMAIL_FUNCTION_SECRET:latest
+            RESEND_EMAIL_ADDRESS=ALERTS_SENDER_EMAIL:latest
+          env_vars_update_strategy: overwrite
+          secrets_update_strategy: overwrite
+          flags: "--port=8000"
