+++ edge-functions/alerts/template.ts
